+++ changelog.md
@@ -2,11 +2,14 @@
 
 ## Upcoming
 * Moved `dimensions` and `biomes` whitelist to a `world_whitelist` object
+* Fixed Slot not checking for world whitelist correctly
+* Fixed attribute names breaking them
+* Game no longer crashes when InsaneLib is missing
 
 ## Alpha 3.0.1
 * Revamped Difficulty modifier
-The difficulty object is now `difficulty_modifier` containing a modifier for difficulty, for local difficulty and if should affect the max value only.
-Weight Modifier now uses the `difficulty` object (`operation`, `easy`, `normal`, `hard`)
+The difficulty object is now `difficulty_modifier` containing a modifier for difficulty, for local difficulty and if should affect the max value only.  
+Weight Modifier now uses the `difficulty` object (`operation`, `easy`, `normal`, `hard`)  
 Chance and Attributes make use of the new `difficulty_modifier`
 * `override_vanilla` has been renamed to `override`
 * Chance is now required with a value between 0 and 1 instead of 0 and 100